refactor(db): move git repository store ownership
